requirements/requirements-all.txt
sqlalchemy>=2.0.0
logmuse
peppy==0.35.6
ubiquerg>=0.6.2
coloredlogs>=15.0.1
pytest-mock
pydantic<2.0
psycopg2-binary
